refactor: OpenAPI와 CLI 도움말 정의 분리
