+++ src/pages/openapiSpecActions.ts
@@ -1,401 +1,8 @@
 /**
  * OpenAI Actions용 축약 OpenAPI 스펙
  */
-import { ACTION_QUERY_ACTIONS } from '../api/actionsProxy.js';
+import { ACTION_QUERY_PARAMETERS } from './openapiSpecActionParameters.js';
 import { OPENAPI_COMPONENTS } from './openapiSpecComponents.js';
-const ACTION_QUERY_PARAMETERS = [
-  {
-    name: 'action',
-    in: 'query',
-    required: true,
-    description: '실행할 액션 이름입니다. 사용 가능한 값은 enum 목록을 참고하세요.',
-    schema: {
-      type: 'string',
-      enum: ACTION_QUERY_ACTIONS,
-    },
-  },
-  {
-    name: 'q',
-    in: 'query',
-    required: false,
-    description: '다이소 제품 검색어',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'query',
-    in: 'query',
-    required: false,
-    description: '세븐일레븐 상품 검색어',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'itemCode',
-    in: 'query',
-    required: false,
-    description: 'GS25 상품 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'productId',
-    in: 'query',
-    required: false,
-    description:
-      '제품 ID 또는 상세 조회용 path 파라미터. 상품명만 알면 먼저 daisoSearchProducts 또는 /api/daiso/products로 productId를 확인하세요.',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'keyword',
-    in: 'query',
-    required: false,
-    description: '공통 키워드 파라미터',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'location',
-    in: 'query',
-    required: false,
-    description: '주변 장소 검색 지역/역/주소 키워드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'category',
-    in: 'query',
-    required: false,
-    description: '주변 장소 카테고리(restaurant, cafe, food, dessert, all)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'services',
-    in: 'query',
-    required: false,
-    description: '통합 비교 서비스 목록(daiso,gs25,seveneleven,emart24)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'type',
-    in: 'query',
-    required: false,
-    description: '개발자 요청 유형(bug, improvement, feature, docs, other)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'severity',
-    in: 'query',
-    required: false,
-    description: '개발자 요청 영향도(low, medium, high, critical)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'title',
-    in: 'query',
-    required: false,
-    description: '개발자 요청 제목',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'description',
-    in: 'query',
-    required: false,
-    description: '개발자 요청 상세 설명',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'toolName',
-    in: 'query',
-    required: false,
-    description: '문제가 발생했거나 개선이 필요한 MCP 도구 이름',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'reproduction',
-    in: 'query',
-    required: false,
-    description: '오류 재현 방법',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'expected',
-    in: 'query',
-    required: false,
-    description: '기대 결과',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'actual',
-    in: 'query',
-    required: false,
-    description: '실제 결과',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'source',
-    in: 'query',
-    required: false,
-    description: '요청 출처(mcp, chatgpt, claude, cli, api, other)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'sido',
-    in: 'query',
-    required: false,
-    description: '다이소 시/도',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'gugun',
-    in: 'query',
-    required: false,
-    description: '다이소 구/군',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'dong',
-    in: 'query',
-    required: false,
-    description: '다이소 동',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'area',
-    in: 'query',
-    required: false,
-    description: '롯데마트 지역',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'area1',
-    in: 'query',
-    required: false,
-    description: '이마트24 1차 지역',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'area2',
-    in: 'query',
-    required: false,
-    description: '이마트24 2차 지역',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'areaCode',
-    in: 'query',
-    required: false,
-    description: '메가박스 지역 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'brandVariant',
-    in: 'query',
-    required: false,
-    description: '롯데마트 브랜드 변형',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'lat',
-    in: 'query',
-    required: false,
-    description: '위도',
-    schema: { type: 'number', format: 'float' },
-  },
-  {
-    name: 'lng',
-    in: 'query',
-    required: false,
-    description: '경도',
-    schema: { type: 'number', format: 'float' },
-  },
-  {
-    name: 'page',
-    in: 'query',
-    required: false,
-    description: '페이지 번호',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'pageIdx',
-    in: 'query',
-    required: false,
-    description: '올리브영 페이지 번호',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'pageLimit',
-    in: 'query',
-    required: false,
-    description: '롯데마트 추가 페이지 수',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'pageSize',
-    in: 'query',
-    required: false,
-    description: '페이지 크기',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'size',
-    in: 'query',
-    required: false,
-    description: '페이지 크기 또는 검색 크기',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'offset',
-    in: 'query',
-    required: false,
-    description: 'CU 검색 오프셋',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'limit',
-    in: 'query',
-    required: false,
-    description: '최대 결과 수',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'storeLimit',
-    in: 'query',
-    required: false,
-    description: '매장 결과 제한 수',
-    schema: { type: 'integer' },
-  },
-  {
-    name: 'storeCode',
-    in: 'query',
-    required: false,
-    description:
-      '매장 코드. 다이소 진열 위치는 /api/daiso/inventory 응답의 storeInventory.stores[].storeCode에서 확인하세요.',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'storeName',
-    in: 'query',
-    required: false,
-    description:
-      '매장명. 롯데마트 상품 조회에서 storeCode가 없을 때 사용하며, 모르면 먼저 lottemartFindNearbyStores 또는 /api/lottemart/stores로 확인하세요.',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'storeKeyword',
-    in: 'query',
-    required: false,
-    description: '매장 키워드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'playDate',
-    in: 'query',
-    required: false,
-    description: '상영일(YYYYMMDD)',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'regionCode',
-    in: 'query',
-    required: false,
-    description: 'CGV 지역 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'theaterCode',
-    in: 'query',
-    required: false,
-    description: 'CGV 극장 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'theaterId',
-    in: 'query',
-    required: false,
-    description: '메가박스/롯데시네마 지점 ID',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'movieCode',
-    in: 'query',
-    required: false,
-    description: 'CGV 영화 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'movieId',
-    in: 'query',
-    required: false,
-    description: '메가박스/롯데시네마 영화 ID',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'pluCd',
-    in: 'query',
-    required: false,
-    description: '이마트24 상품 PLU 코드',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'bizNoArr',
-    in: 'query',
-    required: false,
-    description: '이마트24 사업자번호 목록',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'label',
-    in: 'query',
-    required: false,
-    description: '세븐일레븐 인기 검색어 라벨',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'sort',
-    in: 'query',
-    required: false,
-    description: '정렬 기준',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'searchSort',
-    in: 'query',
-    required: false,
-    description: 'CU 검색 정렬 기준',
-    schema: { type: 'string' },
-  },
-  {
-    name: 'service24h',
-    in: 'query',
-    required: false,
-    description: '이마트24 24시간 필터',
-    schema: { type: 'boolean' },
-  },
-  {
-    name: 'includeSoldOut',
-    in: 'query',
-    required: false,
-    description: '올리브영 품절 포함 여부',
-    schema: { type: 'boolean' },
-  },
-  {
-    name: 'includeIssues',
-    in: 'query',
-    required: false,
-    description: '세븐일레븐 이달의 행사지 포함 여부',
-    schema: { type: 'boolean' },
-  },
-  {
-    name: 'includeExhibition',
-    in: 'query',
-    required: false,
-    description: '세븐일레븐 전시상품 포함 여부',
-    schema: { type: 'boolean' },
-  },
-  {
-    name: 'timeoutMs',
-    in: 'query',
-    required: false,
-    description: '세븐일레븐 inventory 타임아웃',
-    schema: { type: 'integer' },
-  },
-] as const;
 
 export function generateOpenApiSpec(baseUrl: string): object {
   return {
